closeClient: rm clients from channels

+++ srcs/Server.cpp
@@ -118,7 +118,7 @@ void Server::runIRC()
 		else if (status == 0)
 			continue;
 			
-		for (int i = static_cast<int>(_pollFds.size()) - 1; i >= 0; --i)
+		for (int i = _pollFds.size() - 1; i >= 0; --i)
 		{
 			if (_pollFds[i].revents & POLLIN)
 			{
